buyer_app/requirements.txt
kivy>=2.2.0
xrpl-py>=2.4.0
# pyjnius>=1.4.0  # Only needed for Android - uncommented during buildozer build
